move to helpers, buffered -> bufferedIntersection
