+++ src/util/to-title-case.js
@@ -1,0 +1,9 @@
+const toTitleCase = function(string) {
+  if (typeof string !== 'string') {
+    return string;
+  }
+
+  return string.replace(/./, (w) => w.toUpperCase());
+};
+
+export default toTitleCase;
